requirements.txt
irc
spotipy
python-dotenv
requests
discord.py
psutil
flask
flask-cors
gunicorn
psycopg2-binary